fix: cppcheck duplicated
